TSO-7 Add media content to community.html
Add <figure> elements in each <section> element of community.html. .gif
images were used. A Discord iframe of the Obsidian Discord server was
added--which was taken from the frame source in Obsidian's website. An
image of the Obsidian forum from the website was included as well. All
resources from Obsidian has its corresponding <cite> element in the
<figcaption> tag. The <ul> elements contain aria-label attributes for
its list headings.