TSO-4 Update <iframe> element in compare.html
Remove any text or markup inside the <iframe> element as its content
model is 'nothing'. Remove frameborder attribute of <iframe> as it is
deprecated and should be set in CSS instead. Update the title attribute
of the <iframe> element for accessibility purposes.

Add a comment in the <head> element to set a reminder for frameborder in
CSS.