TSO-5 Update <iframe> element and fix <p> in using-obsidian.html
Update the <iframe> element in line 121. Remove frameborder attribute
and added a comment at the <head> element as a reminder to include it in
the external CSS later. Update the title attribute with an appropriate
title.

The <p> element in line 101 should not have a block element (in this
case, the <ol> tag) nested inside it. The <ol> element was moved outside
and the <p> element is now closed at line 103.