TSO-10 Modify <form> element in contact.html
Restructure the form elements inside the <form>. Inserted each form
control inside a <p> tag. Specified for attributes in the <label>
elements as well as the name and id attributes for each form control.

Restructured the <select> element as well. Updated the first <option> to
be selected, disabled, and hidden. Wrapped the <option> elements in an
<optgroup> as well.

Updated the <form> attributes with the method and action specified.
Temporarily set the action to the same webpage "./contact.html".